Review fixes: extract shared highlightCode, add caching, fix HTML semantics
- Extract duplicated highlightCode function to @shared/highlighting/highlightCode.ts,
  used by both markdownRenderer.ts and HistoryItem.ts
- Lazy-initialize MarkdownIt instance (deferred until History tab renders)
- Cache rendered markdown HTML per instruction to avoid re-parsing on every
  Lit update cycle (toggle, highlight index changes)
- Fix invalid HTML: change <span> to <div> for instruction history-value
  since rendered markdown produces block-level elements
- Make .md extension check case-insensitive in memory file opener
